requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.1.0=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
astroid=2.5.6=pypi_0
async-generator=1.10=pypi_0
attrs=21.2.0=pypi_0
autopep8=1.5.7=pypi_0
backcall=0.2.0=pypi_0
base58=2.1.0=pypi_0
bleach=3.3.0=pypi_0
blinker=1.4=pypi_0
ca-certificates=2021.4.13=haa95532_1
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py39haa95532_0
cffi=1.14.5=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
colorama=0.4.4=pypi_0
cycler=0.10.0=pypi_0
decorator=5.0.9=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.3=pypi_0
gitdb=4.0.7=pypi_0
gitpython=3.1.17=pypi_0
idna=2.10=pypi_0
ipykernel=5.5.5=pypi_0
ipython=7.23.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
isort=5.8.0=pypi_0
jedi=0.18.0=pypi_0
jinja2=3.0.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-core=4.7.1=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=pypi_0
lazy-object-proxy=1.6.0=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.4.2=pypi_0
matplotlib-inline=0.1.2=pypi_0
mccabe=0.6.1=pypi_0
mistune=0.8.4=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.3=pypi_0
nest-asyncio=1.5.1=pypi_0
notebook=6.4.0=pypi_0
numpy=1.20.3=pypi_0
openssl=1.1.1k=h2bbff1b_0
packaging=20.9=pypi_0
pandas=1.2.4=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.2.0=pypi_0
pip=21.0.1=py39haa95532_0
prometheus-client=0.10.1=pypi_0
prompt-toolkit=3.0.18=pypi_0
protobuf=3.17.0=pypi_0
pycodestyle=2.7.0=pypi_0
pycparser=2.20=pypi_0
pydeck=0.6.2=pypi_0
pygments=2.9.0=pypi_0
pylint=2.8.2=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pyserial=3.5=py39haa95532_0
python=3.9.4=h6244533_0
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pywin32=300=pypi_0
pywinpty=1.1.1=pypi_0
pyzmq=22.0.3=pypi_0
requests=2.25.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=52.0.0=py39haa95532_0
six=1.16.0=pypi_0
smmap=4.0.0=pypi_0
sqlite=3.35.4=h2bbff1b_0
streamlit=0.82.0=pypi_0
terminado=0.10.0=pypi_0
testpath=0.5.0=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.1=pypi_0
tornado=6.1=pypi_0
traitlets=5.0.5=pypi_0
tzdata=2020f=h52ac0ba_0
tzlocal=2.1=pypi_0
urllib3=1.26.4=pypi_0
validators=0.18.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.2=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wincertstore=0.2=py39h2bbff1b_0
wrapt=1.12.1=pypi_0
